chore: Update .goreleaser.yaml to use lowercase OS name
The .goreleaser.yaml file was updated to use the `tolower` function to convert the OS name to lowercase. This ensures compatibility with the results of the `uname` command. Previously, the `lower` function was used, which caused issues on certain systems. This change improves the reliability and portability of the build process.