chore: Update SCP command in release workflow to use absolute path for scp executable
This commit modifies the .github/workflows/release-docs-dev.yaml file to update the SCP command in the release workflow. It replaces the relative path for the scp executable with the absolute path "/usr/bin/scp". This change ensures that the correct scp executable is used for file copying, improving the reliability of the release process.