chore: Update SCP command in release workflow to use SSH_PRIVATE_KEY without removing carriage return
This commit modifies the .github/workflows/release-docs-dev.yaml file to update the SCP command in the release workflow. It changes the way the SSH_PRIVATE_KEY is written to the CERT_PATH file, removing the tr command that was used to remove the carriage return. This change ensures that the SSH_PRIVATE_KEY is written correctly, improving the reliability of the release process.