chore: Update SCP command in release workflow to use single quotes for SSH_PRIVATE_KEY
This commit modifies the .github/workflows/release-docs-dev.yaml file to update the SCP command in the release workflow. It changes the way the SSH_PRIVATE_KEY is written to the CERT_PATH file, using single quotes instead of double quotes. This change ensures that the SSH_PRIVATE_KEY is written correctly, improving the reliability of the release process.